+++ front/nuxt.config.js
@@ -17,8 +17,8 @@ export default {
     link: [{ rel: "icon", type: "image/x-icon", href: "/favicon.ico" }],
   },
   env: {
-    appPublicKey: process.env.APP_PUBLIC_KEY,
-    appPrivateKey: process.env.APP_PRIVATE_KEY,
+    contractAddress: process.env.CONTRACT_ADDRESS,
+    contractNetEndpoint: process.env.ENDPOINT,
   },
   css: [],
   plugins: [],
