+++ package.json
@@ -42,6 +42,7 @@
     "typescript": "^4.5.5"
   },
   "dependencies": {
-    "@openzeppelin/contracts": "^4.4.2"
+    "@openzeppelin/contracts": "^4.4.2",
+    "web3": "^1.7.0"
   }
 }
